Start MITMProxy options and plugins (#3)
* List open dirs in `index.php`, moved from `index.html` with some
not-so-good-lookin PHP code :(

Added initial extension to block outgoing api.telegram POST requests.
This will immediately return a 201 and is on by default. Id like to add
additional logging plus the payload parameters in the POST request.

It makes me think -> we should have a generic POST plugin that'll dump
parameters to logs!

* Update docker-compose.yml

Had a debugging thing going on, nothing to see here.

* Update block-telegram.py

* Quick fix for http/s proxy

* List open dirs in `index.php`, moved from `index.html` with some
not-so-good-lookin PHP code :(

Added initial extension to block outgoing api.telegram POST requests.
This will immediately return a 201 and is on by default. Id like to add
additional logging plus the payload parameters in the POST request.

It makes me think -> we should have a generic POST plugin that'll dump
parameters to logs!

* Update compose to play nicely with generating certs

Co-authored-by: sysgoblin <cbaylissmk2@gmail.com>