Mail interception and other changes (#5)
* List open dirs in `index.php`, moved from `index.html` with some
not-so-good-lookin PHP code :(

Added initial extension to block outgoing api.telegram POST requests.
This will immediately return a 201 and is on by default. Id like to add
additional logging plus the payload parameters in the POST request.

It makes me think -> we should have a generic POST plugin that'll dump
parameters to logs!

* Update docker-compose.yml

Had a debugging thing going on, nothing to see here.

* Update block-telegram.py

* List open dirs in `index.php`, moved from `index.html` with some
not-so-good-lookin PHP code :(

Added initial extension to block outgoing api.telegram POST requests.
This will immediately return a 201 and is on by default. Id like to add
additional logging plus the payload parameters in the POST request.

It makes me think -> we should have a generic POST plugin that'll dump
parameters to logs!

* Update compose to play nicely with generating certs

* Added vscode dir

* Moved mail wrapper and put copy in dockerfile

* Removed ref to phpsendmail

* Removed ref to phpmailsend volume

* Added helper function cos yea

* Added mitmproxy addon refs

* Increase logging, dump to /logs inc sessions

* Basic mail intercept mitmproxy addon

* Dir list function and general improvement

Co-authored-by: zmallen <zma4580@gmail.com>